After code review refactoring
